Added URL params to pickRandomQuestion and removeQuestion

+++ src/functions/RemoveQuestions.js
@@ -2,16 +2,23 @@ const { app } = require('@azure/functions');
 const sql = require('mssql')
 const connString = process.env['dbconn'];
 
+// Will delete a question from the database, given an ID number.
+
+// URL Parameters:
+// - id: The id of the question to delete. If left blank, -1 is used, and no questions are deleted from the database.
+
 app.http('RemoveQuestions', {
     methods: ['GET', 'POST'],
     authLevel: 'anonymous',
     handler: async (request, context) => {
         const pool = await sql.connect(connString);
+        const target = request.query.get('id') || -1;
 
-        // Later on, these parameters will be passed into the API when it is called, but for demo purposes they are set here.
-        var target = -1;
-
-        const data = await pool.request().query("DELETE FROM [dbo].[QuizQuestions] WHERE id = " + target);
+        if (target != -1) {
+            const data = await pool.request().query("DELETE FROM [dbo].[QuizQuestions] WHERE id = " + target);
+        } else {
+            const data = "";
+        }
 
         context.res = {
             body: data
